# Optimize DNS packet parsing performance with aggressive speed improvements