… analysis/optimization_proposal.md
… analysis/working_optimizations.exs
… bench/baseline_extended_bench.exs
… bench/extended_speed_bench.exs
… bench/final_parse_comparison.exs
… bench/lazy_bench.exs
… bench/parse_comparison.exs
… bench/question_focused_baseline.exs
… bench/quick_speed_test.exs
… bench/simple_baseline_bench.exs
… bench/simple_question_baseline.exs
… bench/simple_question_optimized.exs
… bench/speed_bench.exs
… bench/speed_focused_bench.exs
… lib/dns_packet.ex