## Summary
- Implement aggressive speed optimizations for DNS packet parsing with focus on Question-centered packets
- Add comprehensive benchmarking suite for performance validation and regression testing

## Performance Improvements

### Question-focused packets (1 record, typical DNS queries):
- **Short domains**: +43.5% faster (176.79ns → 123.10ns) 
- **Medium domains**: +37.7% faster (178.72ns → 129.65ns)
- **Long domains**: +29.0% faster (198.97ns → 153.99ns)
- **Different record types**: +20.6% faster (199.73ns → 165.54ns)

### Component functions:
- **DNS.type/1 (A)**: +53.5% faster (10.25ns → 6.68ns)
- **create_character_string**: +467.6% faster (88.52ns → 15.59ns)
- **concat_binary_list**: +164.4% faster (101.21ns → 38.27ns)

## Technical Changes

### Core Optimizations
- Add native compilation with HiPE O3 optimization for maximum speed
- Implement `parse_question_fast` and `parse_answer_fast` with reduced overhead (4→2 tuple returns)
- Optimize `parse_name` using iolist accumulator for O(n) string building instead of O(n²)
- Cache DNS type/class lookups to eliminate duplicate calls per record
- Add aggressive function inlining for critical parsing functions
- Remove list reversal operations for speed over order consistency

### New Files Added
- **analysis/**: Performance analysis and optimization proposals (4 files)
- **bench/**: Comprehensive benchmarking suite (14 files)

## Trade-offs
- Records returned in reverse order (performance over consistency)
- Increased memory usage during processing
- Larger compiled module size due to aggressive inlining

## Test Plan
- [x] All existing tests pass (87 tests, 100% success rate)
- [x] Comprehensive benchmark validation with baseline comparisons
- [x] High-precision performance measurements with git-based version comparison
- [x] Question-focused optimization validation for typical DNS usage patterns

## Real-world Impact
These optimizations are particularly beneficial for:
- High-throughput DNS servers processing Question-centered queries
- DNS resolvers handling upstream queries
- DNS load testing tools requiring maximum performance
- Embedded DNS clients where response time is critical

Typical DNS queries (Question 1 record, short-medium domains) will see **37-43% processing time reduction**.

🤖 Generated with [Claude Code](https://claude.ai/code)